Ensure that rotation is smoothed when SmoothLerpToTarget is True.
When SmoothLerpToTarget is true, the target value that we are
interpolating to should be lerpTargetLocalRotation. For example, on line
320 we compute angleDiff assuming the new angle we are going to is
lerpTargetLocalRotation.

However, when performing the actual Quaternion.Slerp computation, we are
currently setting the target as targetRotation. Fix this by setting the
target in the Quaternion.Slerp call to lerpTargetRotation and
lerpTargetLocalRotation.