Exposing occlusion material to address issue #58
Spatial Mapping renders Hot Pink because it uses Resources.Load to load
materials that do not exist under a resources folder. This change will
still require that the user set the occlusion and render materials
directly (or else they'll see hot pink), but now exposes the 'Occlusion
Material' as a property that can be set in the Inspector.