Changed Input Axes to variables to allow for..
non standard unity input configurations.

Due to the unique nature of Hololens. None of the projects I have in the
works use the standard unity input configuration(partially due to trying
to get a bluetooth game controller to work with unity and hololens, no
luck there yet).  The changes submitted here allow the user to make
changes select which axes to use from the ManualCameraControl instead of
having to dig into the input settings.