Spatial Mapping updates (issue #146)
This commit addresses issue #146. The SpatialMappingComponent folder is
no longer needed in the HoloToolkit, because these components
(SpatialMappingRenderer and SpatialMappingCollider) are now built into
Unity. However, providing a basic sample scene is useful to show how to
use the components, so I moved the old test scene into the new
HoloToolit-Examples folder and updated it to use the new built-in Unity
components.