Log error when Sharing binaries are not found in the External folder
Scenario: Developer copies the cloned HoloTookit files into their own
project and External folder gets left behind. Even if they use a unity
package, the Externals folder is not part of that package.

Fix: Log an error when the files are not found in the expected path.
Fixes HoloToolkit-Unity GitHub issue #135