Fixed cursors to offset along gaze direction.
Modified cursors to apply the DistanceFromCollision offset along the
gaze direction rather than the normal of the hit object. Using this
method the cursor stays in line with gaze even with large
DistanceFromCollision values.

Prior to this change if GazeManager.Normal has a large angle of
incidence with the gaze direction the cursor would no line up with gaze,
the higher DistanceFromCollision the worse it would be.  Try out the
ColorPicker example with and without the change to see the improvement.