Support lerping with unscaled time to enable pause menus
This change enables editor input, tagalongs, and the cursor to work
while the time scale is set to 0. Setting the time scale to 0 is a
common practice to implement game "pause" behavior as is done in the
Unity sample project. With this change, tagalong menus will still
function even when the game is paused.