Fix SpatialUnderstandingExample query menu to accept input
The box collider on the menu (required by the tagalong) had a z depth
of 1, and that was blocking raycasts from reaching the UI canvas.
Setting the depth to 0 lets the raycasts reach the buttons on the UI
canvas.

The tab switch buttons at the top were being disabled when their tab
content was being disabled which meant there was no way to switch to a
disabled tab. Leaving the tab switch buttons with their default enabled
value solves this.