Added possibility to lock the objects rotation while dragging it
When u drag a object via HandDraggable and move your head the rotation of the object changes.
At least in my case I don't want that, it somehow feels unnatural, so I implemented a public flag to prevent this behaivour (lock the rotation of the object).
Also I added a Header and Space for grouping the different rotation modes in the inspector (probably change the text there a bit?)
And I grouped the handling of the rotation modes (upright, towards user, lock orientation)  into an if-else if-else if construct. If nothing applies (when IsRotationLocked == true) the rotation of the object isn't changed at all (my desired behaivour).
Any opinions?