Added confirmation dialogues to EnforceEditorSettings
The existing version does not feature a prompt, and modifies the
project configuration without warning. Users with complex or unique
project setups may encounter issues.

Users will now be prompted to change their project settings when
importing the HoloToolkit, or opening a project with the toolkit
already included, to better support external VCS and project
maintainability.