Change OverrideFocusedObjectTest scene to use global listener mechanism
In a recent change, the OverrideFocusedObject concept was removed from FocusManager since it
was redundant with global listeners and modal handlers. This change fixes up the
OverrideFocusedObjectTest scene to use the global listener mechanism.

It also adds a box collider to the text so it can be clicked when not overriding focus and makes the
text easier to understand by counting clicks and saying whether or not focus is currently overridden.