Make ToggleLogic virtual
Wanted to make a customized ToggleInteractive Script which inherits from ToggleInteractive and needed to change the ToggleLogic() mehtod (use a dynamic Unity bool Event). For this I need ToggleLogic() to be virtual and I think this could help others, as well, because in the current form it is not possible to implement a custom toggle logic.