Fix for dragging of rotation handles of BoundBoxRig
Current method of rotation is not consistent; the object is rotated different directions
depending on orientation of object.

This commit adds an option (rotateAroundPivot) which allows the user to rotate an object in a more natural way.

To avoid breaking existing usage, this new rotation is enabled via a bool in BoundingBoxRig