Keyboard Fix for missing keys
There is a bug with the keyboard which causes the first button in each row to not appear or at least be positioned correctly.

I forget the exact reasoning behind it as I made this change locally 6 months ago but I believe it had something to do with gameObjects being inactive if you initialized a keyboard and then hid it.

In any case, it appears to be unecessary overhead for the layout to update at runtime.