Partially removed if-branching from _BORDER_LIGHT
I have followed the documentation on here:

https://shadowmint.gitbooks.io/unity-material-shaders/support/syntax/step.html

Thus I have removed the 2nd branching in the _BORDER__LIGHT-definition that wrote to o.uv.z and o.uv.w and replaced with a custom #define IF at line 122 . Branching in shaders are usually undesirable because of the GPU-architecture (as far as I have read), but sometimes they can be useful in edge-cases. For this case I could have eliminated the if-else ifs in the block above, but on second thought I decided to leave it as to be honest I have not yet spent that much time looking into the core of this shader.