Minor Billboard bug fix
LookRotation wants the two vectors to be othogonal. Giving it the
object's current up vector caused some drift in the z angle in some
cases (since the resulting up vector cannot be what we pass in). Now, we
don't pass any up vector in and let Unity handle it in cases where we
don't want z touched.