Update BoundingBox.cs
Removed a code line in method RefreshTargetBounds() that performed a scale transformation which was correct only in special, simple cases. In general, if the target object, its parent and the bounding box had different scales, the transformation would be incorrect.
The replacement for the removed line, a scale transformation that works in any case, is implemented in method UpdateScaleTransform(), where it is needed and so that the variable name targetBoundsLocalScale does not have to be changed to remain appropriate.