Add scaffolding for diagnostic profile and display
The intent is to provide the layout for what configuration of diagnostic display will be. This does not include any UI elements yet, but still has
the layout there.

Currently supported diagnostic information:

FPS
Memory
CPU

Each diagnostic is only allowed to be turned on or off. Granularity of information will be determined as implementation
for the display is added.