Added _HoverLightRadius for hover light radius
Greetings,
I have added _HoverLightRadius to the shader to allow for finer radius control of the hover light. I have also made the gradient exponential instead of linear, so it now more closely matches the hover light effect elsewhere in the HoloLens OS. 

IMPORTANT:

This also needs the edits I have done in StandardShaderGUI.cs to show in the editor.