Adding support for the .NET scripting background
Placed entirely behind a UNITY_EDITOR check because this particular code is only used from the editor. The UWP .NET libraries don't have the same support for Types class functionality, so rather than trying to update this to support the UWP version of types, we can drop the code in runtime environments.