Updates the default MRTK project settings to use Single Pass Instance Rendering.
As part of working with a number of customers, we've found that many people are using the default Multi Pass rendering mechanism, which is a slow way of rendering.  We've seen in more complex scenarios that this can lead to slowdowns of 5ms per frame.

We should also author a perf doc to recommend setting this value in customers' projects (along with other things like disabling the included visual profiler when doing perf measurements)