Update the MRTK root profile picker to select a reasonable default.
One of the points of customer feedback we've seen internally is that as a new person configuring a scene for the first time, they see a list of ~5 profiles of which they don't know what to pick. We should at least point to a reasonable default (i.e. the default profile), even if we don't want to filter down this list further to avoid breaking advanced user scenarios.