Clean up delivery mechanism documentation.
We currently do not support several of the other mechanisms listed in the DownloadingTheMRTK page, and there's nothing more frustrating when coming upon docs than seeing text that says "coming soon" that isn't helpful.

Our docs should reflect the current state of how to use the product, and when additional support is added for other delivery mechanisms, we should then update the docs to describe those mechanisms.