Fix up a few of the doccomments for the EnumFlags and Edge types
I'm going through the process of looking through our codebase to learn more about everything, and as I'm seeing random issues like this I'm fixing them.

In the case of EnumFlagsAttribute.cs we shouldn't have to go to a link to understand the context.

In the case of Edge.cs it was referencing what appeared to be an older naming of the type.