Standardize on the way we wait for InputSystem validity.
There are multiple ways we wait for InputSystem validity, some of which
check for the current state of the object and some of which use the await directly.

AFAICT it looks like we should have moved to a common helper (i.e. in this PR: https://github.com/Microsoft/MixedRealityToolkit-Unity/pull/3462/files), so this common-izes this action into the base class