Removal of dead code, part N+1
Explanation of the things being removed:

SolverHandlerInspector.cs and ControllerElement.cs:

ControllerElement is dead and no longer used, though it looks like it was used at some point and made it into the solver handler inspector. It could still be useful if someone wanted to make it useful in the future.

QuarternionSmoothed.cs:

Dead and not used.

IMixedRealityPlacementHandler:

Dead and not used.