Workaround for mis-ordered initialization of some input providers.
We've seen a couple of instances now where the MixedRealityInputModule and various providers (i.e. GazeProvider) are getting initialized in non-deterministic fashion. If the GazeProvider starts first, it will raise its source detected, and then when MixedRealityInputModule::ActivateModule is invoked, it re-invokes OnSourceDetected, which tries to add the same pointer to the pointer dictionary.

This throws an exception (bad). We want to update the OnSourceDetected function to only try to add it to the dictionary if it doesn't exist to work around this and unblock a few consumers.

https://github.com/Microsoft/MixedRealityToolkit-Unity/issues/4084