Initial fix for NearInteractionTouchable volumes
Changes NearInteractionTouchable.TouchableSurface to allow for quads or boxes. Quads use the old logic for DistanceToSurface. Boxes use shortest distance to a cuboid in space for DistanceToSurface.

This change may be breaking, as the TouchableSurface enum has changed, meaning that the value that was previously UnityUI is now Box. The UI will still function, but the distance to surface may be strange at the edges. There is a helper in the inspector that will amend this.