Delete the non-functional AttachToController
Per issue: https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4033

It looks like this solver was added but never fully completed (all of the interesting code was commented out). Rather than ship code that doesn't work, we'll delete it. If we're doing a controller or solvers pass, we can add this back at that time.