Fix for the air-tap gesture on HL1 holographic remoting.
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/3909

This issue has been around for a while, and has some interesting history. The underlying cause is a bug in a Unity API (see the code for details), which was fixed at some point in the past, but regressed at some unknown time. With no ETA on when the underlying Unity bug, we're checking in a workaround to our code so that customers can get unblocked.

Note that this scopes the fix to only situations where remoting is active.