Simplify enabling/disabling eye tracking feature support with MRTK, Part 1
There are three steps that you need to take to enable eye tracking, and one of them is flipping a checkbox on the GazeProvider. The interesting thing is that even if this is set to true, the gaze provider will end up defaulting back to HL1 behavior anyway, so there's no harm in having the setting default to true - for folks who have HL1, they will still get HL1 behavior. For folks building on HL2, they will have one less step to take.

Note that this does increase the work for folks who have HL2 and actively want HL1 behavior, however.