Add a default set of profiles that are optimized for getting started with HoloLens2
The default MRTK profile is good at showing off a lot of the features in VR and AR, but has some extra settings that either aren't necessary in AR (teleportation, boundary) or aren't performant (hand mesh visualization). Furthermore, there are also some settings that must be enabled specially in order to use HL2 features (such as Eye Tracking), which should really be set by default.

This copies a subset of the configuration profiles with deltas as follows:

1. Boundary system is disabled.
2. Teleport system is disabled.
3. Spatial awareness system is disabled.
4. Eye tracking provider is added (https://microsoft.github.io/MixedRealityToolkit-Unity/Documentation/EyeTracking/EyeTracking_BasicSetup.html#create-an-eye-gaze-data-provider)
5. Eye simulation is eanbled by default (for input simulation).
6. Hand mesh visualization is disabled.
7. Camera profile settings are set to match, so that visual quality in the editor and device will look similar.