Turning off "Enable Hand Joint Visualization" turns off hand input in Editor
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4268

It turns out that the hand pan interaction script was directly calling into the hand visualization code in order to get joint locations, instead of just calling the hand APIs themselves (which have the exact same data). As I understand, historically the only way of getting this information was through the visualization class. It's been a while that this hasn't been the case, so now it's good to just go straight to the joint data.