Reintroduce the Interpolator utilities back to the MRTK
https://github.com/microsoft/MixedRealityToolkit-Unity/pull/4191

The previous linked change got rid of some code that wasn't referenced anywhere in the MRTK and had no documentation - along with the private->public payload that came back in March, there was a significant portion of dead code that was introduced and this got cleaned up with it. Based on some feedback from customers, they were in fact using these utilities, so we're adding them back with tests to show how things should expect to work.

I'd love to have more coverage on Interpolator.cs, but this would require some minor refactoring (non breaking changes) in order to actually unit test those things (in particular, we'd need to abstract away the usage of time in order to properly test the script)