Add missing namespaces to resolve the lack of NuGet packages building.
The addition of these files without namespaces caused an issue building NuGet packages, because this test was added without a namespace. The asset retargetting step (which is a precursor to building nuget packages) assumes that every single class belongs to the non-global namespace (actually, the MRTK main namespace).

Other followups associated with this issue:

1) Probably some linter that checks that all .cs files actually have a namespace declared.
2) Probably treating warnings from the nuget step/asset retargetting step as error (so that it fails the CI build)