Update the Spatial Awareness docs to provide a guide on how to access meshes
We had some repeated questions about how to do this, so I figured it would be nice just to have this in docs.

Un-obsoletes the GetObserver* functions, because the spatial awareness system doesn't actually have the data provider interface that it references (i.e. it tell you to check to see if it has it, and then you do, and you realize it doesn't have it)