Clean up the the MRTK architecture and packaging documents
I'm in the process of putting up some new MRTK overview docs (i.e. overall architecture, philosophy, that kinda stuff) and when I started that process, I was looking to see what already existed. A lot of this stuff was either not true, aspirational, or no longer accurate. I wanted to ensure that our documents actually reflect the current truth, not the past, not the future, only the current truth.

I bent this rule a little bit with the MRTK_Modularization file (which still contains references to future), though for those, they should be easy enough just to remove the "Upcoming Changes" section once those are done (and they will be done very soon)