Add initial profile documentation.
Our documentation around profiles is sorely lacking. This is the first step toward getting more stuff:

- Have a basic description of what profiles are for.
- Have a basic description of the default profile.
- Have a description of the HoloLens2 profile and how it's different.

This also deserves having some of the rationale/reasoning behind why there's a profile system, but for that we need to get some more chats going to make sure we're on the same page.

This also updates the main getting started page to recommend using the HoloLens2 profile if that's the system they're targetting.