Update the getting started docs to include UWP build setting recommendations.
This is such a common question that I wanted to add some comments in the getting started doc about it. People are trying to configure their target/platform in some cases to the highest number, but in reality the defaults are actually pretty good. The only key part is the requirement to check that 18362 exists in the list, and if not, provides the solution for how to fix it.

Also changes the requirements section up a little to list the requirement link up front, and THEN providing clarification. The reality is that if you're targetting cross-platform with MRTK (which you should be) we shouldn't tell people that "oh actually it's not REALLY needed" and have them run into weird stuff down the line - let's get people to have the same prereqs that we have (which is just the latest set of SDKs) and lead them down the path of success.