Update the UWP build tools to support adding the Gaze Input capability automatically
This change is intended to help make the eye tracking work a little easier to consume. Eye tracking requires the gazeInput capability setup on the Visual Studio project, but Unity currently doesn't support having this functionality in Unity itself. As a result, every single time you rebuild the Unity project, you have to modify your AppX manifest and re-add the gazeInput capability. This is a huge pain for people who want to use eye tracking.

This change gets around this by adding an optional build flag that, when set, will add the capability to the manifest, and this will happen either at:

1) After the Unity build.
2) Before running the appx build.

The function itself is idempotent anyway so it's safe to invoke in both cases. This will also handle cases where the appx solution was built a while back (or was done with the unity tooling instead of our tooling).