Make it so that MixedRealityRaycastHit doesn't throw an exception on every frame.
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4966

When the default raycaster goes through a physics hit and doesn't hit anything, the generated RaycastHit isn't usable. The raycaster returns true when a hit is successful (the only time when the out parameter should be accessed) but otherwise the out parameter shouldn't be used (and it should be null instead)

There's a few things that happened here:

1) lightmapCoord itself is kinda funky when the physics hit doesn't hit anything (if you see the stuff here https://github.com/Unity-Technologies/UnityCsReference/blob/master/Modules/Physics/ScriptBindings/Dynamics.bindings.cs you'll find that it actually tries to do stuff, but the object basically isn't valid to do stuff.) Since RaycastHit is a struct it's not a nullable type, but those objects really aren't safe to use in the case where the physics hit failed.
2) Changed MixedRealityRaycastHit to a class to make it a nullable type (so that we can return a null raycast hit in the case that nothing is hit)
3) Updated the IMixedRealityRaycastProvider to document that the hitInfo will be null when nothing is hit.

Also adds a test to show how this all works end to end (that it doesn't throw, and that it's null in that case).