Animate cursor based on context
As far as I can see, the only cases where we need to show cursor context is when we're doing two hand manipulation and when we're using a bounding box. These changes update the cursor context state depending on ManipulationHandler and BoundingBox.

If anyone has any additional cases where we need to display a different cursor context, let me know.