Port the WorldAnchorManager from the HTK to MRTKv2.
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4509

The WorldAnchorManager provided a convenience wrapper on top of Unity's WorldAnchor/AnchorStore abstractions. This change ports the file mostly as-is into the Experimental namespace, because we haven't yet figured out how best to have end to end tests for this, and because we don't have any documentation yet.

Note that there have been several changes here from the HTK version:

1) Some functions that were protected/virtual have been made private, for the sake of going out with the minimal API surface needed to use the World Anchor Manager. This doesn't mean that we they will be private forever, just that we don't want to create a massive surface and want to extend it as-needed.
2) Removed usage of the Singleton<> template - it's up to the consumer how they choose to use this (they can do their own singleton if they want)