Add MRTK architecture documentation, part 1 of N.
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4852
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4779

One of the sore spots of our documentation is explaining what all is in the MRTK, how things connect, why WHY things were written in the way they were written. I wrote up an internal doc a couple weeks back that inventoried all of the code and features in the MRTK. Some folks found this to be helpful in providing a high level grounding as they dove deeper into the toolkit, so I'm publishing this in hopes that it will help everyone else.

Long term, this documentation should be kept up to date and provide not just how things work, but why things work.