Enforce compiler warnings as errors in the MRTK.
It is critically important to ensure that the MRTK builds without warnings and errors. We have consumers that have strict "warnings as errors."

Over the past month, we've seen a number of changes go in with compiler warnings - these never should have been checked in in the first place.

This change adds a C# compiler flag to will treat all compiler warnings as errors.

Note that this DOES have a ramification on a subset of consumers that currently pull directly from mrtk_development (and mrtk_release, once this change goes out as part of GA). There's significant benefit to ensuring that our code compiles without warnings or errors.