Ensure that asset retargeting fails on non-zero Unity exit codes.
We've already seen a number of cases where we have failed to push NuGet packages, because of classes that were introduced with no namespaces. These show up as an error in the asset retargeting step, but because we don't return the error code that Unity gives us, the rest of the build moves on (and then later shows a warning saying that "no nuget packages were pushed out")

This updates the build powershell script so that any break that would cause us to stop producing a nuget package would actually cause a build failure, rather than just a warning and slap on the wrist.