Fix the broken play mode tests (PointerEventsTests)
These pointer tests aren't working in CI, despite working in the editor itself.

This was driving me crazy for a while, and from what I can see, this is due to the frame rate tick of the game itself vs as the test code is running - it looks like in batch mode especially, frames might not actually be passing between test lines, so things that assume that an update has occurred, or that a certain amount of drag events have occurred.

I was able to see this just by doing things like increasing the numSteps to 26 (or some larger than 1 thing) to increase the chances that frames were actually passing.

I wanted to keep that number relatively low, so I did a few other things instead:

1) Wait for a fixed update after each manipulation (i.e. gesture, end of move), and set the target framerate to match the physics tick, so as to increase the chance that we're actually going to get the frame to pass.
2) Remove the exact checking for drag events, because depending on the system, more or less events can pass between each test line.

I also removed the prefab in favor of just dynamically newing up a Cube at the same location and size.