Fix the broken Standalone build.
https://github.com/microsoft/MixedRealityToolkit-Unity/pull/5018 broke the standalone build, but we didn't roughly today.

Here's what happened.

TouchScreenKeyboard is only documented to be useable on the WSA/iOS/Android build options. It turns out that the "visible" property just doesn't exist on the object when build against Standalone targets. The code will build fine in editor but when you actually try to build the player, it doesn't work.

So to that end, I've wrapped the rest of the code in platform specific wrappers (to target only the places where TouchScreenKeyboard is documented to work)

I only caught this because I was making asset retargeting actually fail the build if it failed:

https://github.com/microsoft/MixedRealityToolkit-Unity/pull/5103

I'm going to make a change after this to ensure that player builds will actually fail the build if they fail, instead of just silently succeeding.