Update the build player step to fail the build if the player fails to build.
It turns out right now, if the player build fails, the build can silently succeed. This has happened with the Standalone build, and we haven't caught this because we have depended on the BuildAppX stage in order to catch that the player never succeeded. Of course, with Standalone builds, there is no AppX stage, so failures there just don't get caught.

This change does a couple of things.
1) Makes it so that a failure code returned by the unity build step (i.e. build failure or exception) will actually fail the build, instead of silently succeeding.
2) Changes it so that we will display the unity log in the same step in the case of failure (if we didn't change the "print whatever logs" stages, the failure would have prevented the next steps from going, and we wouldn't see the logs.)

Note that this works with this PR:
https://github.com/microsoft/MixedRealityToolkit-Unity/pull/5103

To make sure that we'll fail the build and still get useful logging information.