Update the build pipeline to first build standalone, in order to avoid infinitely hanging play mode tests.
It turns out that if there's a build error, running play mode tests will hang indefinitely in batch mode (this is because it shows a dialog saying "hey, make sure your stuff builds"). Previously this wasn't an issue because the test step happened after everything else, but now that we moved the test step up to be the first thing, this is now problematic.

Building standalone is the fastest of all of the flavors (because it doesn't have to build an AppX), so we're moving this to the top.